Valid or Invalid input tied to record value change and view in etymological symbology
